Separate sanitization class
